Added support for CERES, size now is an integer

+++ af_io_test.h
@@ -24,15 +24,15 @@ double* af_read_hyperslab(hid_t file, char*dataset_name, int x_offset, int y_off
 hsize_t* af_read_size(hid_t file, char* dataset_name);
 
 //Instrument data retrieval functions
-double* get_misr_rad(hid_t file, char* camera_angle, char* resolution, char* radiance, double* size);
-double* get_misr_lat(hid_t file, char* resolution, double* size);
-double* get_misr_long(hid_t file, char* resolution, double* size);
-double* get_modis_rad(hid_t file, char* resolution, char* d_name, double* size);
-double* get_modis_lat(hid_t file, char* resolution, char* d_name, double* size);
-double* get_modis_long(hid_t file, char* resolution, char* d_name, double* size);
-double* get_ceres_rad(hid_t file, char* camera, char* d_name, double* size);
-double* get_ceres_lat(hid_t file, char* camera, char* d_name, double* size);
-double* get_ceres_long(hid_t file, char* camera, char* d_name, double* size);
+double* get_misr_rad(hid_t file, char* camera_angle, char* resolution, char* radiance, int* size);
+double* get_misr_lat(hid_t file, char* resolution, int* size);
+double* get_misr_long(hid_t file, char* resolution, int* size);
+double* get_modis_rad(hid_t file, char* resolution, char* d_name, int* size);
+double* get_modis_lat(hid_t file, char* resolution, char* d_name, int* size);
+double* get_modis_long(hid_t file, char* resolution, char* d_name, int* size);
+double* get_ceres_rad(hid_t file, char* camera, char* d_name, int* size);
+double* get_ceres_lat(hid_t file, char* camera, char* d_name, int* size);
+double* get_ceres_long(hid_t file, char* camera, char* d_name, int* size);
 
 //Helper functions
 void concat_by_sep(char** source, const char** w, char* sep, size_t length, int arr_size);
